pyproject.toml
[tool.poetry]
name = "ws2811-mqtt"
version = "1.2.1"
description = "This package makes it possible to control the ws2811 type lights with a homeassistant mqtt instance"
authors = ["Guillaume Richard <lord0gnome@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "ws2811_mqtt"}]

[tool.poetry.scripts]
ws2811-mqtt = 'ws2811_mqtt:main'

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^2.1.0"
rpi-ws281x = "^5.0.0"
adafruit-circuitpython-neopixel = "^6.3.12"
adafruit-blinka = "^8.50.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
